feat(compile): --windows-subsystem override + tier-3 Apple link dedup + pointer-width portability (v0.5.1125) (#4685)
- --windows-subsystem auto|console|windows flag + perry.toml [windows]
  subsystem fallback; precedence CLI > perry.toml > import heuristic.
  Lets a GUI app that doesn't import perry/ui suppress the console window.
- Tier-3 Apple (tvOS/watchOS) link dedup: localize std panic/unwind shims
  (__rust_drop_panic, rust_begin_unwind, ...) alongside allocator shims so
  -Zbuild-std runtime+stdlib don't collide with ld64.lld duplicate symbols.
- Pointer-width portability: compare alloc size as u64 against 1u64<<34 in
  gc/barrier, gc/copying, timer, typedarray; widen tm_gmtoff cast in date.

+++ crates/perry-runtime/src/date.rs
@@ -191,7 +191,7 @@ fn timestamp_to_local_components(secs: i64) -> (i32, u32, u32, u32, u32, u32, i6
         let hour = tm.tm_hour as u32;
         let minute = tm.tm_min as u32;
         let second = tm.tm_sec as u32;
-        let tz_offset = tm.tm_gmtoff;
+        let tz_offset = tm.tm_gmtoff as i64;
         (year, month, day, hour, minute, second, tz_offset)
     }
 }
